rep_orchestrator_cli::orchestrator

Trait Client

pub trait Client: Send + Sync {
    // Required methods
    async fn update_status(
        &self,
        status: Status,
        exit_status: Option<ExitStatus>,
        message: Option<String>,
    ) -> Result<()>;
    async fn generate_upload_urls(&self) -> Result<UploadUrls>;
    async fn upload_to_signed_url(&self, url: &str, body: Vec<u8>) -> Result<()>;

    // Provided method
    async fn update_error_status(&self, error: CliError) -> Result<()> { ... }
}

Required Methods

async fn update_status( &self, status: Status, exit_status: Option<ExitStatus>, message: Option<String>, ) -> Result<()>

Update the [Status] of the current test execution, with an optional message to write to the database.

If there was a subprocess associated with this status update, its ExitStatus may be included for additional context.

async fn generate_upload_urls(&self) -> Result<UploadUrls>

Request signed URLs from the orchestrator for uploading the log file and output zip associated with the current test execution.

async fn upload_to_signed_url(&self, url: &str, body: Vec<u8>) -> Result<()>

PUT body to a previously-issued signed upload URL.

Provided Methods

async fn update_error_status(&self, error: CliError) -> Result<()>

Updates the status of the current test execution with context from the provided CliError.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.