rep_orchestrator_cli::context

Trait CliContext

pub trait CliContext {
    type OrchestratorClient: Client;
    type KubeClient: Client;

    // Required methods
    fn orchestrator_client(&self) -> &Self::OrchestratorClient;
    fn kube_client(&self) -> &Self::KubeClient;

    // Provided method
    async fn run_shell(
        &self,
        cmd: &mut Command,
        next_execution_status: Status,
    ) -> Result<(), CliError> { ... }
}

Required Associated Types

type OrchestratorClient: Client

type KubeClient: Client

Required Methods

fn orchestrator_client(&self) -> &Self::OrchestratorClient

fn kube_client(&self) -> &Self::KubeClient

Provided Methods

async fn run_shell( &self, cmd: &mut Command, next_execution_status: Status, ) -> Result<(), CliError>

Run a shell Command to completion.

If it exits with a successful exit code, Self::OrchestratorClient will post an update moving the associated test execution to next_execution_status.

If it fails, Self::OrchestratorClient will be used to update the status of the associated test execution to either a [Status::Unrunnable] or [Status::Failed] state depending on whether the failed command was associated with invoking user scripts or setup scripts.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.